requirements.txt
PyQt5==5.15.10
Pillow==10.1.0
